fix console prop

+++ bootstrap/src/main/java/com/udacity/bootstrap/entities/Dog.java
@@ -6,7 +6,7 @@ import javax.persistence.*;
 public class Dog {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     @Column
     private String name;
     @Column
@@ -14,11 +14,11 @@ public class Dog {
     @Column
     private String origin;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -46,7 +46,7 @@ public class Dog {
         this.origin = origin;
     }
 
-    public Dog(Integer id, String name, String breed, String origin) {
+    public Dog(Long id, String name, String breed, String origin) {
         this.id = id;
         this.name = name;
         this.breed = breed;
